+++ flexagent-core/src/main/java/org/flexagent/core/exception/FlexAgentException.java
@@ -1,0 +1,12 @@
+package org.flexagent.core.exception;
+
+public class FlexAgentException extends RuntimeException {
+    
+    public FlexAgentException(String message) {
+        super(message);
+    }
+
+    public FlexAgentException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}